rename the package from server to rater
