+++ pkg/daemon/server/service/rater/options.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package server
+package rater
 
 type options struct {
 	// Number of workers working on collecting counts of processed messages.
