+++ pkg/daemon/server/service/rater/uniq_str_list_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package server
+package rater
 
 import (
 	"testing"
